+++ tests/Api/CalendarApiCest.php
@@ -19,7 +19,7 @@ class CalendarApiCest
 
     public function _before(ApiTester $I): void
     {
-        $this->region = $I->createRegion();
+        $this->region = $I->createRegion(fillMailbox: false);
 
         $this->user = $I->createFoodsharer();
         $I->addRegionMember($this->region['id'], $this->user['id']);