+++ tests/Api/UserApiCest.php
@@ -17,7 +17,7 @@ class UserApiCest
 {
     private $user;
     private $userOrga;
-    private $store;
+    private $region;
     private $faker;
 
     private const EMAIL = 'email';
@@ -33,12 +33,9 @@ class UserApiCest
         $I->addRegionMember($group['id'], $this->user['id']);
         $I->addRegionMember($group['id'], $this->userOrga['id']);
 
-        $region = $I->createRegion();
-        $I->addRegionMember($region['id'], $this->user['id']);
-        $I->addRegionMember($region['id'], $this->userOrga['id']);
-
-        $this->store = $I->createStore($region['id']);
-        $I->addStoreTeam($this->store['id'], $this->user['id']);
+        $this->region = $I->createRegion(fillMailbox: false);
+        $I->addRegionMember($this->region['id'], $this->user['id']);
+        $I->addRegionMember($this->region['id'], $this->userOrga['id']);
 
         $this->faker = Factory::create('de_DE');
     }
@@ -463,9 +460,12 @@ class UserApiCest
 
     public function canDeleteUser(ApiTester $I): void
     {
+        $store = $I->createStore($this->region['id']);
+        $I->addStoreTeam($store['id'], $this->user['id']);
+
         // add user to a pickup slots
-        $I->addPicker($this->store['id'], $this->user['id']);
-        $I->addPicker($this->store['id'], $this->user['id'], ['confirmed' => 0]);
+        $I->addPicker($store['id'], $this->user['id']);
+        $I->addPicker($store['id'], $this->user['id'], ['confirmed' => 0]);
 
         // delete user
         $I->login($this->user[self::EMAIL]);
@@ -474,12 +474,12 @@ class UserApiCest
 
         // check that the user is not in the team anymore and that no future slots are assigned to the user
         $I->dontSeeInDatabase('fs_betrieb_team', [
-            'betrieb_id' => $this->store['id'],
+            'betrieb_id' => $store['id'],
             'foodsaver_id' => $this->user['id']
         ]);
         $I->dontSeeInDatabase('fs_abholer', [
             'foodsaver_id' => $this->user['id'],
-            'betrieb_id' => $this->store['id'],
+            'betrieb_id' => $store['id'],
             'date >' => Carbon::now()->format('Y-m-d H:i:s')
         ]);
     }