+++ tests/Unit/EventGatewayTest.php
@@ -34,9 +34,9 @@ class EventGatewayTest extends Unit
 
         $this->regionGateway = $this->tester->get(RegionGateway::class);
         $this->foodsaver = $this->tester->createFoodsaver();
-        $this->region = $this->tester->createRegion('God');
+        $this->region = $this->tester->createRegion('God', fillMailbox: false);
         $this->tester->addRegionMember($this->region['id'], $this->foodsaver['id']);
-        $this->childRegion = $this->tester->createRegion('Jesus', ['parent_id' => $this->region['id']]);
+        $this->childRegion = $this->tester->createRegion('Jesus', ['parent_id' => $this->region['id']], false);
     }
 
     public function testAddLocation(): void