+++ tests/Unit/RegularPickupGatewayTest.php
@@ -21,7 +21,7 @@ class RegularPickupGatewayTest extends Unit
     public function _before()
     {
         $this->gateway = $this->tester->get(RegularPickupGateway::class);
-        $this->region = $this->tester->createRegion();
+        $this->region = $this->tester->createRegion(fillMailbox: false);
         $this->store = $this->tester->createStore($this->region['id']);
     }